fixed lint errors
