+++ app/docs/api/page.tsx
@@ -1,36 +1,43 @@
-import { createSwaggerSpec } from 'next-swagger-doc';
-import dynamic from 'next/dynamic';
-import 'swagger-ui-react/swagger-ui.css';
+'use client';
 
-const SwaggerUI = dynamic(
-  () => import('swagger-ui-react'),
-);
+// import { createSwaggerSpec } from 'next-swagger-doc';
+// import dynamic from 'next/dynamic';
+// import 'swagger-ui-react/swagger-ui.css';
 
-export const metadata = {
-  title: 'API Documentation - Markket',
-  description: 'API documentation for Markket Next.ts application'
-};
-
-async function generateSpec() {
-  return createSwaggerSpec({
-    apiFolder: 'app/api',
-    definition: {
-      openapi: '3.0.0',
-      info: {
-        title: 'MarkketNext API',
-        version: '1.0',
-      },
-    },
-  });
-}
+// const SwaggerUI = dynamic(() => import('swagger-ui-react'), {
+//   ssr: false,
+//   loading: () => (
+//     <div className="flex items-center justify-center h-screen">
+//       <div className="loader"></div>
+//     </div>
+//   ),
+// });
 
 
-export default async function ApiDoc() {
-  const spec = await generateSpec();
+// // async function generateSpec() {
+// //   return createSwaggerSpec({
+// //     apiFolder: 'app/api',
+// //     definition: {
+// //       openapi: '3.0.0',
+// //       info: {
+// //         title: 'MarkketNext API',
+// //         version: '1.0',
+// //       },
+// //     },
+// //   });
+// // }
 
+
+export default function ApiDoc() {
+  // const spec = await generateSpec();
+
+  // return (
+  //   <div className="api-docs">
+  //     <SwaggerUI spec={spec} />
+  //   </div>
+  // );
   return (
-    <div className="api-docs">
-      <SwaggerUI spec={spec} />
-    </div>
+    <div>swagger</div>
   );
 };
+
