+++ public/robots.txt
@@ -1,0 +1,3 @@
+User-agent: *
+Disallow: /dashboard/
+
